Deprecate source_common
